only allow placing towers on grass

+++ Enemy.cs
@@ -36,7 +36,7 @@ internal class Enemy
 
     public float jumpCooldown = 0;
     public float collisionRadius = 0;
-    public float goldValue = 10;
+    public int goldValue = 10;
 
 
     public Rectangle GetRect()
